Print time spent in compilation queue in JIT performance vlogs
When the -Xjit:verbose=compilePerformance option is enabled, record the
time spent in the compilation queue in JIT verbose logs for successful
compilations and AOT cache hits (as "... queueTime=NNNus"). This time
represents the total compilation request latency which is a useful JIT
performance metric. The high-resolution (microsecond) timestamp of when
the compilation is queued is stored in the MethodToBeCompiled struct.

Signed-off-by: Alexey Khrabrov <khrabrov@cs.toronto.edu>